cleanup capture fragment
