+++ Sensory/src/main/java/com/google/android/sensory/fhir_data/PhotoCaptureViewHolderFactory.kt
@@ -35,11 +35,8 @@ import com.google.android.fhir.datacapture.views.QuestionnaireViewItem
 import com.google.android.fhir.datacapture.views.factories.QuestionnaireItemViewHolderDelegate
 import com.google.android.fhir.datacapture.views.factories.QuestionnaireItemViewHolderFactory
 import com.google.android.sensing.SensingEngine
-import com.google.android.sensing.capture.CaptureFragment
-import com.google.android.sensing.capture.CaptureSettings
-import com.google.android.sensing.model.CaptureInfo
-import com.google.android.sensing.model.CaptureType
-import com.google.android.sensing.model.InternalSensorType
+import com.google.android.sensing.capture.sample.CameraCaptureFragment
+import com.google.android.sensing.capture.sensors.CameraCaptureRequest
 import com.google.android.sensory.InstructionsFragment
 import com.google.android.sensory.R
 import com.google.android.sensory.SensingApplication
@@ -116,10 +113,10 @@ object PhotoCaptureViewHolderFactory :
         val parentFragmentsChildFragmentManager =
           view.findFragment<QuestionnaireFragment>().parentFragmentManager
         parentFragmentsChildFragmentManager.setFragmentResultListener(
-          CaptureFragment.TAG,
+          CameraCaptureFragment.TAG,
           context
         ) { _, result ->
-          if (result.getBoolean(CaptureFragment.CAPTURED)) {
+          if (result.getBoolean(CameraCaptureFragment.CAPTURED)) {
             // Following condition arises when user presses back button from CaptureFragment and the
             // fragment is removed from backstack by the BackPressCallback defined in
             // ScreenerFragment
@@ -127,7 +124,7 @@ object PhotoCaptureViewHolderFactory :
               parentFragmentsChildFragmentManager.popBackStack()
             }
 
-            val captureId = result.getString(CaptureFragment.CAPTURED_ID, "")
+            val captureId = result.getString(CameraCaptureFragment.CAPTURED_ID, "")
             context.lifecycleScope.launch {
               sensingEngine.getCaptureInfo(captureId).resourceInfoList.forEach {
                 val answer =
@@ -172,29 +169,20 @@ object PhotoCaptureViewHolderFactory :
             context
               .getSharedPreferences(SensingApplication.SHARED_PREFS_KEY, Context.MODE_PRIVATE)
               .getString(SensingApplication.CURRENT_PATIENT_ID, null)!!
-          val captureId = questionnaireViewItem.answers.firstOrNull()?.valueCoding?.code
-          val captureFragment =
-            CaptureFragment().apply {
-              setCaptureInfo(
-                CaptureInfo(
+          val cameraCaptureFragment =
+            CameraCaptureFragment().apply {
+              setCaptureRequest(
+                CameraCaptureRequest.ImageRequest(
                   externalIdentifier = fhirPatientId,
-                  captureType = CaptureType.IMAGE,
-                  captureFolder =
+                  outputFolder =
                     "Sensory_${SensingApplication.APP_VERSION}/Participant_$fhirPatientId/$QUESTION_TITLE",
-                  captureSettings =
-                    CaptureSettings(
-                      fileTypeMap = mapOf(InternalSensorType.CAMERA to "jpeg"),
-                      metaDataTypeMap = mapOf(InternalSensorType.CAMERA to "tsv"),
-                      captureTitle = QUESTION_TITLE
-                    ),
-                  recapture = captureId != null,
-                  captureId = captureId,
+                  outputTitle = QUESTION_TITLE,
                 )
               )
             }
           parentFragmentsChildFragmentManager
             .beginTransaction()
-            .add(R.id.screener_container, captureFragment, CaptureFragment.TAG)
+            .add(R.id.screener_container, cameraCaptureFragment, CameraCaptureFragment.TAG)
             .addToBackStack(null)
             .commit()
         }
