+++ sensing/src/main/java/com/google/android/sensing/db/impl/DbTypeConverters.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 Google LLC
+ * Copyright 2023-2024 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,8 +17,6 @@
 package com.google.android.sensing.db.impl
 
 import androidx.room.TypeConverter
-import com.google.android.sensing.capture.CaptureSettings
-import com.google.gson.Gson
 import java.time.Instant
 
 internal object DbTypeConverters {
@@ -27,10 +25,4 @@ internal object DbTypeConverters {
   @JvmStatic
   @TypeConverter
   fun longToInstant(value: Long?): Instant? = value?.let { Instant.ofEpochMilli(it) }
-
-  @JvmStatic @TypeConverter fun toString(value: CaptureSettings) = Gson().toJson(value)
-
-  @JvmStatic
-  @TypeConverter
-  fun toCaptureSettings(value: String) = Gson().fromJson(value, CaptureSettings::class.java)
 }
