+++ sensing/src/main/res/layout/fragment_image_stream.xml
