+++ sensing/src/main/res/layout/layout_capture_instructions.xml
@@ -19,7 +19,7 @@
     >
 
     <TextView
-            android:id="@+id/ppg_instruction"
+            android:id="@+id/camera_request_instruction"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
             android:layout_weight="1"
@@ -31,7 +31,7 @@
 
     <RelativeLayout android:layout_width="100dp" android:layout_height="100dp">
       <com.google.android.material.progressindicator.CircularProgressIndicator
-                android:id="@+id/ppg_progress"
+                android:id="@+id/timer_progress"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 app:indicatorSize="90dp"
@@ -44,7 +44,7 @@
                 app:trackColor="@android:color/darker_gray"
             />
       <TextView
-                android:id="@+id/record_timer"
+                android:id="@+id/timer_text_view"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:text="00:00"
